Revert "Remove blurb about HTTP redirects" (#6)
This reverts commit c9df87ff9482e6cbfac971b2355950111c86c1d6.

Reasoning for revert:
The description of the reverted commit stated that HTTP redirects are
not followed by URLClassLoader, but it looks like they in fact are
followed.
The test that was added by the reverted commit was not properly implemented,
since the fake HTTP endpoint was not initialized properly. Therefore, the
fake HTTP redirect never got triggered. The test succeeded because nothing
got loaded from the fake non-redirect URL, so the class wasn't found. When
the fake HTTP endpoint was properly initialized using Spark.awaitInitialization,
then the redirect was triggered, and the class did get loaded.